do not define command for mixnet
